etracted utterance class
